Create fictional_landing_page.html
In this lab, you will build a simple Landing Page for a fictional product, service, or game. The focus is on using semantic HTML, clean structure, and user-focused content.

Fulfil the user stories below and get all the tests to pass to complete the lab.

User Stories:

You should have a DOCTYPE declaration.

You should have an html element with lang set to en.

You should have a head element with a meta charset of utf-8 and a title element with the text Landing Page.

You should have a header element with:

An h1 element with the name of the product, game, or service

A nav element with an unordered list of links to three sections: Features, Testimonials, and Get Started

You should have a main element with three sections:

A section with id="features" that includes:

An h2 element titled Features

A list (ordered or unordered) with at least three unique features

A section with id="testimonials" that includes:

An h2 titled Testimonials

At least two blockquotes, each with a short quote and a citation using <cite>

A section with id="get-started" that includes:

An h2 titled Get Started

A short paragraph inviting the user to sign up or try the product

A link styled as a button (use <a href="#">) with text like "Join Now" or "Play Free"

You should have a footer element with a brief copyright.