Create pepo_bio_page.html
In this lab, you'll build a Character Bio Page for a fictional (or real) character. This could be a CS2 agent, a meme legend like Pepo, an anime character, a game boss, or even yourself.

Fulfil the user stories below and get all the tests to pass to complete the lab.

User Stories:

You should have a DOCTYPE declaration.

You should have an html element with lang set to en.

You should have a head element with:

A meta charset of utf-8

A title with the character’s name

You should have a header with:

An h1 element with the character’s name

A paragraph summarising who they are

You should have a main element with three sections:

A section with id="bio" that includes:

An h2 titled Biography

Two or more paragraphs describing the character’s backstory or personality

A section with id="stats" that includes:

An h2 titled Stats

A table showing at least three stats (e.g. Accuracy, Win Rate, Favourite Weapon)

Use a <thead>, <tbody>, and <tfoot> to summarise total or average

A section with id="gallery" that includes:

An h2 titled Gallery

Two figure elements, each with:

An image and a figcaption

You should have a footer with a quote from the character or a final line of your choice.